# openapi/daf-openapi.yaml
x-commons:
  ratelimit-headers: &ratelimit-headers
    X-RateLimit-Limit:
      $ref: '#/components/headers/X-RateLimit-Limit'
    X-RateLimit-Remaining:
      $ref: '#/components/headers/X-RateLimit-Remaining'
    X-RateLimit-Reset:
      $ref: '#/components/headers/X-RateLimit-Reset'
  common-responses: &common-responses
    '400':
      $ref: '#/components/responses/400BadRequest'
    '429':
      $ref: '#/components/responses/429TooManyRequests'
    '503':
      $ref: '#/components/responses/503ServiceUnavailable'
    default:
      $ref: '#/components/responses/default'

openapi: 3.0.0
info:
  title: Pdnd API
  version: 0.1.0
  contact:
    email: alessandro@teamdigitale.governo.it
    name: Alessandro Ercolani
  description: |
    #### Documentazione
    Il servizio si propone di esporre tramite openapi3 alcune API del PDND - Daf
    #### Note
    Il servizio richiede autenticazione tramite Bearer o Basic Auth previa registrazione sul portale
     https://dataportal.daf.teamdigitale.governo.it , e va' usato rispettando gli
    header di throttling esposti in conformita' alle Linee Guida del Modello di interoperabilita'.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - description: Development server
    url: http://localhost:8080/pdnd-openapi
  - description: Test server
    url: https://api.daf.teamdigitale.it/pdnd-openapi
paths:
  /status:
    get:
      description: |
        Return JWT token
      operationId: app.status
      responses:
        <<: *common-responses
        '200':
          content:
            'application/json':
              schema:
                type: object
                properties:
                  status:
                    type: string
          description: |
            Current status of api.
          headers:
            <<: *ratelimit-headers
      summary: Return status from server
      tags:
      - Status
  /jwt:
    get:
      description: |
        Return JWT token
      operationId: app.get_token
      security:
        - basic: []
      responses:
        <<: *common-responses
        '200':
          content:
            'application/json':
              schema:
                type: object
                properties:
                  jwt:
                    type: string
          description: |
            JWT token.
          headers:
            <<: *ratelimit-headers
      summary: Return pdnd jwt token
      tags:
      - Login
  /public/search:
    post:
      description: |
        Return JWT token
      operationId: app.public_data_search
      security:
        - jwt: []
      requestBody:
          content:
            application/json:
              schema:
                 x-body-name: filters
                 $ref: '#/components/schemas/Filters'
          #description: filter to use in search
          #required: true
      responses:
        <<: *common-responses
        '200':
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
          description: |
            JWT token.
          headers:
            <<: *ratelimit-headers
      summary: Return pdnd jwt token
      tags:
      - Search
  /dataset/save:
    post:
      description: |
        Save dataset into PDND. Remember only user with particula privileges can use
        this API.
      operationId: app.dataset_save
      security:
        - jwt: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              x-body-name: file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                theme:
                  type: string
                subtheme:
                  type: string
                org:
                  type: string
                user:
                  type: string
                description:
                  type: string
              required:
                - file
          #description: filter to use in search
          #required: true
      responses:
        <<: *common-responses
        '200':
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
          description: |
            JWT token.
          headers:
            <<: *ratelimit-headers
      summary: Return pdnd jwt token
      tags:
      - Datasets
  /dataset/save_to_test:
    post:
      description: |
        Save dataset into 
      operationId: app.dataset_save_test
      security:
        - jwt: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              x-body-name: file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
              required:
                - file
          #description: filter to use in search
          #required: true
      responses:
        <<: *common-responses
        '200':
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
          description: |
            JWT token.
          headers:
            <<: *ratelimit-headers
      summary: Return pdnd jwt token
      tags:
      - Datasets
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: app.forward_token
    basic:
      type: http
      scheme: basic
      x-basicInfoFunc: app.basic_auth
  headers:
    X-RateLimit-Limit:
      $ref: 'https://teamdigitale.github.io/openapi/headers/v3.yaml#/X-RateLimit-Limit'
    X-RateLimit-Remaining:
      $ref: 'https://teamdigitale.github.io/openapi/headers/v3.yaml#/X-RateLimit-Remaining'
    X-RateLimit-Reset:
      $ref: 'https://teamdigitale.github.io/openapi/headers/v3.yaml#/X-RateLimit-Reset'
    Retry-After:
      $ref: 'https://teamdigitale.github.io/openapi/headers/v3.yaml#/Retry-After'

  responses:
    # Predefined error codes for this API
    400BadRequest:
      $ref: 'https://teamdigitale.github.io/openapi/responses/v3.yaml#/400BadRequest'
    404NotFound:
      $ref: 'https://teamdigitale.github.io/openapi/responses/v3.yaml#/404NotFound'
    429TooManyRequests:
      $ref: 'https://teamdigitale.github.io/openapi/responses/v3.yaml#/429TooManyRequests'
    503ServiceUnavailable:
      $ref: 'https://teamdigitale.github.io/openapi/responses/v3.yaml#/503ServiceUnavailable'
    default:
      $ref: 'https://teamdigitale.github.io/openapi/responses/v3.yaml#/default'
  schemas:
    Filters:
      type: object
      properties:
        text:
          type: string
        index:
          type: array
          items:
            type: string
        org:
          type: array
          items:
            type: string
        theme:
          type: array
          items:
            type: string
        date:
          type: string
        status:
          type: array
          items:
            type: string
        owner:
          type: string
        order:
          type: string
        sharedWithMe:
          type: boolean
    SearchResult:
      type: object
      properties:
        type:
          type: string
          description: type of the source
        source:
          type: string
          description: the source
        match:
          type: string
          description: the match result
    